requirements.txt
Django==2.2.5
djangorestframework=3.10.3
drf-yasg[validation]=1.16.1
gunicorn==19.9.0
packaging=19.1
psycopg2-binary==2.8.3
pyyaml=5.1.2
uritemplate=3.0.0
whitenoise==4.1.3